Implement card preview in chat (#61)
Allow argument formatting in addMessage.
Implements the ability to pass arguments to addMessage and have them be
formatted within the message text. These arguments can be strings or
arbitrary JSON objects.
The first implementation client-side is detecting that a card was
embedded in the message. The client then renders the card name that when
hovered over will display the card in the image preview.
Convert card messages to use formatter.
Convert game messages to use formatter.
Update a few places using card.label instead of the full card.