Update stealth targeting requirements
A character with stealth may no longer be chosen as a target for
stealth. It is also now possible to choose a stealth target that does
not have an icon for the current challenge. This is useful for cards
that have an ability triggered when they bypass a character by stealth
(such as Wolves of the North).

Fixes #54.