Fix dragging cards from dead/discard to play.
The pipeline click handler added to Game.playCard was returning if it
couldn't find the card within the player's hand. This method is also
called when a card is dragged into the play area. By returning early,
cards dragged from the dead and discard piles would disappear instead of
be put into play.