Save and restore previous card selections.
When a select card prompt comes up during another select card prompt,
the original selections will be restored once the newer prompt has been
completed. The main use case for this seems to be selecting attackers
and defenders during a challenge, and deciding to set the strength of
one of the participants in the middle of selection.