Move active plot can play check into player.
Previously whether a card was allowed to be played by the active plot
was done in Game.playCard. The opponent's plot was also checked.
However, there are currently no cards that can prevent an opponent from
playing certain kinds of cards. If any such cards are introduced in the
future, all of the opponents cards should be asked, not just their
active plot.