Update Player.playCard to take a card directly.
Previously, the playCard method took a card ID and optional source list
to look up the card object. With the introduction of moveCard, it is no
longer necessary to pass around the source list. Thus it's better if
playCard doesn't need to look up the card each time when it usually has
been looked up already.