Update styling of tokens on horizontal cards.
Previously, the tokens on horizontal cards were being displayed
vertically, leading to them being cut off if there were more than 2 of
them. This flips the orientation for to be horizontal which accomodates
3 tokens.

Additionally, if there are more than 3 tokens on a card, a
`.many-counters` class is applied. This class ensures that when there
are 4 or more tokens that they bunch up into a square, rather than two
unequal rows/columns.