Send strength stats for 0 strength characters.
Previously, characters with 0 base strength (e.g. Lancel Lannister)
would not display their strength if it were increased via abilities,
attachments or using the /strength command. This was due to the strength
not being sent if the base strength was a falsy value. Now it explicitly
sends it unless the base strength is explicitly null, which is the case
on non-character cards.