Fix missing claim and reserve in player stats.
Previously, only plots that had a 'When revealed' ability were calling
`Player.revealPlot`. However, that method is where the claim and reserve
are set on the player. Thus if the plot didn't have a when revealed,
claim and reserve were being displayed as zero.

Additionally, claim was being read from the plot incorrectly.
`PlotCard.claim` is always undefined. This has been changed to call the
`getClaim` method.