Remove the inPlay flag from cards.
Previously, the inPlay flag was being used to guard against events
firing for cards that were not in a play area. Since events are now only
bound when a card enters a play area, this check isn't necessary
anymore.

In some cases, inPlay was being used to check whether it was permissible
to target a card in the play area. These usages have been replaced with
an explicit check that the card's location is 'play area'.