Introduce class to manage card actions.
The CardAction class wraps the responsibility of a few common use cases
for card actions. It makes sure that the player executing the action has
permission to do the action (either controls the associated card or the
action allows any player to take it), ensures that the limit of uses for
the action has not been reached, ensures that the card's action has not
been disabled through the card being blanked, and ensures that the
action is permitted in the current phase.

If the action's handler returns false, then the action is considered not
to have been completed and the attempt to execute it does not count
against the limit.