Allow event self handlers when leaving play
Some cards need to react to being killed, sacrificed or discarded.
Because event listeners are removed when a card leaves play, these
events must be fired prior to the card being moved to an out-of-play
game location.

This change ensures that the onCharacterKilled and onSacrificed events
are raised prior to the onCardLeftPlay event.