Support for conditional + recalculating effects
A condition function may be sent when creating an effect to determine
whether an effect should be immediately or delayed until the condition
is met. In addition, a isStateDependent flag has been introduced on
effect implementations to indicate that they are continuously
recalculated. The recalculation happens everytime a server command has
been executed.