Add event window to manage event resolution order
Now when an event is fired, it queues a game step that manages the
ordering of each sub-event fired. The order per the rules is interrupts
that can cancel the condition, forced interrupts, interrupts, the
triggering condition, forced reactions, and then reactions. By putting
it in the pipeline like this, each step (e.g. all interrupts to a
character being killed) can be fully resolved before moving on to the
next window.