Allow reactions to set their title.
Previously, all prompting reactions/interrupts used the same title
asking whether the player wanted to trigger the character. For saving
interrupts, it is necessary to list which character is being saved.
Prompted triggered abilities can now take a title function in their
properties to do this. The title function takes a triggered ability
context object which includes the event information, and should return a
string to be used to prompt the player.