Always do multi-select for challenges
If Jousting Contest was the active plot, the challenger
limit was set to 1. This caused the select prompt to send a single card
object instead of the array that was expected, causing a crash. This
change implements a new flag on the select card prompt that tells it to
act like a multi-select prompt even if the limit is 1.