Ensure onBeginRound event is fired.
The previous change introducing event ordering did so by queuing an
event resolver onto the game pipeline. Therefore the order of queuing
the onBeginRound event and the rest of the game steps matters now.
Previous to this change, reducers were no longer reseting because the
onBeginRound event was being queued after all of the round steps,
thereby never executing.