Fix dynamic and state restoring effects.
Previously, effects that were dynamicly calculated or restored original
state upon being unapplied were using the card object as the key on the
context object. Javascript silently converts these to strings using
`toString`, meaning if such an effect were applied to multiple cards,
they would all use the same key and overwrite each others state.

As a result, cards like Lord of the Crossing which apply their effect to
multiple cards were setting the strength properly, then when unapplied
would correctly reduce the strength of the first card, delete the state
from the context object, and then apply an undefined strength modifier to
the remainder of the cards. Since the remaining cards had `NaN` as a
strength, the resulting challenge strength would be 0.