Fix cards with self-matching persistent effects
Previously the ability to match cards in hand was implemented. This
introduced a bug where cards that did not use a matching function (i.e.
they targetted themselves or another specific card) could have their
effect applied multiple times because the target location was not
enforced. In the case of Robert Baratheon, the effect was applied just
before he entered play (while still in hand), then again when the
onCardEnterPlay event was fired. Then when strength was recalculated for
the card, it was unapplied twice, leading to a NaN strength modifier.