Add support for arbitrary card piles
It is now possible to define more card piles in additional to the built
in ones. Currently, new piles are defined for when a card is removed
from the game ('out of game'), and for scheme plots used by Rains of
Castamere ('scheme plots'). Cards that were removed from the game
previously now explicitly use the moveCard method.