Ensure nested events are not cancelled
Previously, if an interrupt ended up cancelling an event, any events
that may have been triggered during that cancellation would be skipped
because they are queued onto the EventWindow's pipeline.

For example, if Jory Cassel's ability activated when a character would
be killed, the sacrifice and left play events for Jory would not be
emitted. Thus, any cards that react to Jory's sacrifice (e.g. WotN
Catelyn) would not trigger.