Fix For the Watch!
Prior to implementing event windows, events were totally synchronous, so
For the Watch's handler was being called prior to
challenge.initiateChallenge() being called. Since events are handled via
the pipeline now, when the handler is finally fire the # of challenges
listed is 1 instead of 0. See Lord of the Crossing for similar
off-by-one numbering.

Additionally, the onAttackersDeclared event has now been nested inside
of the onChallenge event. This should make handling of these events more
consistent.