Convert 'cannot win' into an effect.
Previously, For the Watch! was using event handlers to set the cannot
win challenge flag. However, cards like Forgotten Plans could
potentially blank FtW so it has been converted to an effect to support
this.

Additionally, unnecessary conditions in ChallengeFlow have been removed.
They were there previously when a winner was being set in all
circumstances. The winner is no longer set if the cannot win flag is in
effect, so these checks are no longer necessary.