Ensure player-order prompts use current order
Previously, if a game step that prompted in first player order (such as
action windows) was queued, it would use the order that existed at the
time it was queues. So, for example, the action window at the beginning
of the challenges phase was queued before a first player was chosen.
This made it always prompt the player that created the game, even if the
other player was the first player. Additionally, in subsequent rounds
the player that was first the previous round would be prompted first
even if they were no longer first player.