Remove checking other cards for 'can play'
Previously, when a card would be played, it would ask permission from
every other card the player had in play to see if it was allowed to be
played. This was only being used by Marching Orders which has now been
converted to an effect. This simplifies the `canPlay` method to only
concern itself with whether play requirements for that specific card
have been met.