Reapply effects when card changes controller
Previously, if a card was taken control of, any persistent effects
generated by that card would not be reapplied under the new controller.
So for example, if a player took control of a card with a plot modifier
such as the +3 from The Arbor, the new controller would not receive the
+3 gold bonus. Additionally, if the card was not dynamic or conditional,
the effect applied to the opponent would persist even after the
controller changed.

Now when a card has changed controllers, it will immediately recalculate
its effects for the new controller.