Fix double strength problem w/ take control
Previously, take control would manually push the card onto the
cardsInPlay array and then cards like Yoren and Euron would call
moveCard to update their location to be 'play area'. Unfortunately, this
was causing the card to exist twice in the cardsInPlay, causing its
strength to be doubled in the dominance phase if it were standing.