Allow event handlers to cancel the event.
In order to enforce the restriction that a card is in play when being
killed, it is now possible to cancel the event within the normal event
handler. Example: Valar is selected as a plot and Jory Cassel + another
unique Stark character are in play. Jory uses his ability to sacrifice
himself to save the other Stark character, and ends up in the discard
pile. Jory is not killed even though he would have been killed normally.