Ensure plot card area is equally spaced.
Previously, the plot card area was centered. This worked fine when there
were only two possible plot piles for either player. Since Castamere
schemes now can show up there, it was centering 5 piles, so one player's
used pile would be across the mid-line of the board. This change ensures
that all of a players plot piles are on their own side of the board.