UI for Castamere schemes and 'out of game' cards. (#364)
* Use `Player.moveCard` to manage plot placement

Previously, the flow of plots from plot deck to active plot to revealed
plots was managed manually. This change converts this flow to use the
`Player.moveCard` method for more consistency and allows Castamere to
move cards into the 'out of game' pile when a scheme leaves play.

* Add support for arbitrary card piles

It is now possible to define more card piles in additional to the built
in ones. Currently, new piles are defined for when a card is removed
from the game ('out of game'), and for scheme plots used by Rains of
Castamere ('scheme plots'). Cards that were removed from the game
previously now explicitly use the moveCard method.

* Display additional card piles in the UI.

When an additional pile has at least 1 card in it, it will be displayed
in the UI. Placement is determined by the 'area' property for the pile.

* Ensure plot card area is equally spaced.

Previously, the plot card area was centered. This worked fine when there
were only two possible plot piles for either player. Since Castamere
schemes now can show up there, it was centering 5 piles, so one player's
used pile would be across the mid-line of the board. This change ensures
that all of a players plot piles are on their own side of the board.