Don't display counters / tokens when facedown (#419)
Previously, if the effects engine caused a card to gain icons during the
setup phase, the icons would appear on the facedown card, potentially
revealing information to the opponent. For example, if the Shadow Tower
Mason was placed along with 3 Night's Watch locations, the intrigue and
military icons gained would be visible to the opponent even if the card
was face down.