Introduce onCardsDiscarded event.
When one or more cards are being discarded from a single location, an
onCardsDiscarded event will be raised for the group of cards. Then, each
card in that group will raise the previously implemented onCardDiscarded
event. The signatures of these events have been changes to take a
consolidated params object instead of individual parameters. This is
because certain abilities need to replace the values involved (e.g.
Tywin LoCR changes which card is discarded).

Because these multiple card discards are wrapped in events,
discardFromDraw and discardAtRandom now take a callback function to be
called once the discard has occurred.