add automatic parsing of deck title/faction/agenda to the deck builder
Complete ThronesDB-like textual decks can now be pasted into the "Cards"
section of the deck builder. If a deck "header" --- made of deck title,
faction, agenda, possibly with multiple banners in the case of "Alliance" ---
is detected, it will be used to pre-populate the values of the various deck
builder dropdown selectors.

Fixes #633