Allow click to activate actions to be undone.
After converting reducer cards to the effects API, it was no longer
possible to cancel the ability by standing the card. Now, for any click
to activate action, clicking the card again will cancel the ability and
unpay the costs associated with the action. Currently this is only
implemented on the kneelSelf cost but can be extended to other costs in
the future should the need arise.