Prevent reactions from firing from wrong location.
Previously, if an interrupt or reaction ended up removing another card
from play (e.g. Harrenhal kills Littlefinger when he enters play), then
the player would still be prompted to trigger the removed card's
abilities. Now reactions check that the card is still in the proper area
to be eligible to activate.

Fixes #640.