Simplify and extract the plot reveal process.
The previous plot reveal process consisted of an onPlotFlip event (which
is being used inappropriately by cards like Old Nan), an onPlotRevealed
event for each plot revealed, then an onPlotRevealCompleted event. This
combination of events was needed because card reactions resolve after
all plot "when revealed" abilities have been resolved.

Now, onPlotFlip still exists, but the rest of the reveal process is
wrapped in an onPlotsRevealed event which lists an array of all plots
being revealed. "When revealed" abilities will be changed to use a
onPlotsWhenRevealed event in conjuction with the ability window inside
of this event. This is necessary for plot duplicators like Varys's
Riddle which initiated the "when revealed" ability without it counting
as revealing that plot.

Because there is a single onPlotsRevealed event wrapping the entire
flow, interrupts and reactions can fire at the correct time.

Additionally, by extracting the reveal process to its own gamestep,
revealing plots outside of the plot phase such as with Rains of
Castamere will be more consistent for cards like Caswell's Keep.