Don't raise onClaimApplied event with Aemon.
The onClaimApplied event represents claim being applied during or for a
challenge so that cards like Mirri Maz Duur and The Seastone Chair can
react to them. Aemon's ability applies claim outside of a challenge, so
it should not raise the event to trigger these abilities. Thus, the
event is now only raised during the challenge flow.