Add a deck search prompt.
Implements a `promptForDeckSearch` method on `Game` to simplify the deck
searching process. The new prompt asks the player to choose a card based
on the number of cards and card condition matcher, then calls an
onSelect callback when they've chosen. Whether they choose a card or
cancel the prompt, the player's draw deck is automatically shuffled.