Add playable location check for all playing types.
Previously, marshalling had a mechanism to allow locations other than
the current player's hand to be valid. That mechanism is now generalized
for playing events and ambushing cards. By default, marshal, playing,
and ambushing are valid from the player's hand. Effects may add new
valid locations.

Both actions and reactions no longer check that events defined by the
ability API are in hand. Instead, they check that the event is in a
playable location for events.