Protect against loops when removing duplicates.
Battle of the Blackwater could trigger an infinite loop if the character
that has the dupe had been taken control from the opponent.

When removing attachments, if the dupe's owner is not the same as the
attachment's owner an infinite loop could occur. There's no scenario
under the current rules where this should be able to happen but we
should still protect against it.