Guarantee a player object when serializing.
Previously, if a spectator was leaving in the middle of a selection
prompt, an error would occur because their spectator object had been
removed from the list of players. Thus, the serializers were unable to
check whether or not a card was selected / selectable by that user.

Now, when state is generated for a player with no name / a player not in
the spectator list, an anonymous spectator object is generated to ensure
that an object is sent during serialization. This opens to the door -
but does not fully implement - to allowing non-logged in users to
spectate on games.