Allow actions to be restricted per card.
It is now possible to restrict an action from being taken on a card
based on restrictions placed on the card. As abilities are activated,
they are placed on a stack on `Game` to keep track of the current
ability's source card. When a game action is applied to a card (e.g.
being killed, discarded, knelt, etc), the card checks through the list
of restrictions to see if that particular action coming from the source
card is allowed. This provides the necessary infrastructure to implement
both immunity abilities as well as conditional 'cannot' effects (e.g.
Ygritte).

To make checking these restrictions more convenient, a `applyGameAction`
method is provided on `Game` that will only apply an action to eligible
cards.