Ensure legacy events placed in discard.
Previously, legacy event cards were being placed in discard as part of
their handler. However, if the effects of the event were cancelled, the
event would not be placed in discard. Now, placing in discard is
considered part of the cost as it is with events defined using the
ability API.