Fix simultaneous event window crash
Previously, if a kill was canceled during a simultaneous kill window, it
would continue to try to fire events for that card. However, the card's
event had already been removed from the window, leading it to try and
open an ability window for an undefined / null event object.

Fixes THRONETEKI-SX
Fixes THRONETEKI-T0
Fixes THRONETEKI-SV
Fixes THRONETEKI-T2