Remove pulse effect from selectable cards. (#907)
With the addition of greying out unselectable cards, the pulse effect on
selectable cards is less necessary. The animation is not very
performant, so dropping it should help client side rendering
performance. The glowing shadow effect for selectable cards is still
retained to make select prompts clear in IE versions that do not support
the grey scale effect.