Fix Lord of the Crossing strength calculation. (#905)
Attacker strength was being announced before the challenge had actually
been initiated. Thus cards dependent on the current challenge number
like Lord of the Crossing were not applying their effects prior to the
strength being announced. This lead to situations where the strength in
the message was one challenge behind where it should be, but the final
strength calculation would be done correctly.