Gaining faction affiliation loses neutrality. (#929)
Per the rules, cards that "are not affiliated with any faction" are
considered neutral. Thus, a neutral card that has gained an affiliation
via cards like Ward or Sworn to the Watch are no longer considered
neutral. This distinction is mildly important for cards like Tithe which
require kneeling a neutral card as a cost.