Prompt for only one copy of max-limited abilities.
If the player has two copies of an event card that has a maximum in
hand, the prompt they see will only display a single copy of that card.
This should help prevent confusion that could be caused by showing both
copies in the prompt, but then not getting a chance to activate the
second one due to the maximum having been reached.