Vengeance for Elia shouldn't transfer power (#943)
Vengeance for Elia changes who claim is applied to but not who is
considered to win the challenge. Thus, if it is played on a power
challenge, Vengeance will cause power to be transfered from the winner
of the challenge to themself, thus no power is gained or lost by either
player.