Allow abilities to trigger when they enter play.
Previously, if an ability triggering brought a card into play or into
hand that should be playable, the new card wouldn't trigger because the
appropriate event listeners weren't bound when the original event
occurred. Now, when a card is moved to a location it would listen to
events, it registers its abilities with the Game object, which will add
the abilities to any prompts that are already open. This allows
interactions such as triggering Rains of Castamere to play Wardens of
the West, or allowing an event card drawn mid-reaction to be playable if
that ability window is still open.