Kneeling Winterfell is an effect, not a cost.
Previously, Winterfell was not kneeling because the `cost` property was
typoed as `costs` (plural). However, in the case of Winterfell, kneeling
is not a cost for the ability; it is a pre-then effect. Thus, if
Winterfell is canceled through cards like Treachery, it will not kneel.
The implementation now reflects this.