Properly handle unique attachments on opponents.
Previously, if the opponent placed a unique attachment on one of your
cards and you attempted to play the same attachment, it would improperly
be placed as a duplicate on top of the opponent's copy. Additionally,
the opponent could place multiple copies of that attachment instead of
it automatically duplicating the original.