Convert Core Stannis to an effect.
Moves the logic for the stand phase off of player and into the
StandingPhase class. Core Stannis is now implemented by adding a
'standing phase restriction' to each player that filters out character
cards from automatically standing.

Additionally, the player is now required to select 2 cards to stand if
they have more than 2 cards. Previously it was implemented as 'up to 2
cards' but standing during the standing phase is mandatory.

Finally, if the player has less than the stand maximum, all cards will
be stood automatically with no player selection.