Implement The Red Wedding. (#996)
* Allow abilities to trigger for other players.

Previously, reactions / interrupts for a card would only prompt the
card's controller. With the introduction of The Red Wedding, it is
necessary to prompt other players as well. A `player` property has been
added to reactions / interrupts that specifies a function that returns
the player that should be prompted. If the property is not specified, it
defaults to prompting the card controller as before.

* Implement The Red Wedding.