Prevent premature kills from burn effects.
Previously, the order burn effects were applied was determined by the
order of players joining the game plus the order that the burn effect
came into play. This lead to bugs where a card both buffed by one card
and burned by another was killed because the burn effect happened first.
Now, the `killByStrength` effect is explicitly ordered to be applied
after other effects so that cards aren't killed inappropriately.